% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelSTR.R
\name{createVariables}
\alias{createVariables}
\title{Create additional measurements to add to the loaded data frame.}
\usage{
createVariables(df, zscore)
}
\arguments{
\item{df}{The dataframe loaded from a text file.}

\item{zscore}{The zscore used to determine whether a STR locus is variable or
not.}
}
\value{
Returns a dataframe with added columns \code{size}, \code{size.mean}, \code{size.sd}
, \code{size.zscore}, \code{motif.totalcount}, \code{variable}, \code{variable.total},
\code{variable.percent}
}
\description{
A function that takes in a dataframe and zscore specified by the user and
adds additional measurements (\code{size}, \code{size.mean}, \code{size.sd}, \code{size.zscore},
\code{motif.totalcount}, \code{variable}, \code{variable.total}, \code{variable.percent}) to the
dataframe. The purpose of this function is to label a STR locus as either
variable or not using a user defined zscore input parameter. For example,
if a user enters a zscore of 0.2, then a STR locus with size greater than 0.2
standard deviations of the mean motif size, will be lablled variable.
}
\examples{
# Update dataframe with additional columns and labeled loci
data("ShortTandemRepeatsLoci")
mydf <- createVariables(ShortTandemRepeatsLoci, 0.1)


}
\references{
R Core Team (2020). R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna,
Austria. URL https://www.R-project.org/.
}
