% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMotifsDistribution}
\alias{plotMotifsDistribution}
\title{Plots the distribution of the specified motifs.}
\usage{
plotMotifsDistribution(df, motifsList, count = TRUE)
}
\arguments{
\item{df}{The dataframe created using labelSTR.R containing labeled STR loci
and measurement variables.}

\item{motifsList}{A list of motifs to graph out.}

\item{count}{If TRUE uses count for y axis Else uses density for y axis.}
}
\value{
Plots out a density or count distribution graph of the specified
motifs.
}
\description{
A function that takes a dataframe containing labeled STR loci and measure
variables, a list of motifs and plots out a distribution graph of motif
size with either count or density on the y axis.
}
\examples{
# Plot out the distribution size of motifs "AGG", "AC"
plotMotifsDistribution(df, c("AGG", "AC"))

}
\references{
R Core Team (2020). R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna,
Austria. URL https://www.R-project.org/.

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag
New York, 2016.
}
