% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTopNCommonMotifs}
\alias{plotTopNCommonMotifs}
\title{Plots the top N most common motifs from the dataframe.}
\usage{
plotTopNCommonMotifs(df, num, stacked = FALSE)
}
\arguments{
\item{df}{The dataframe created using labelSTR.R containing labeled STR loci
and measurement variables.}

\item{num}{Specifies top how many motifs to graph.}

\item{stacked}{If TRUE uses stacked bar plot Else uses a dodge bar plot.}
}
\value{
Plots out a bar plot showing the top N most common motifs and
the proportion of variable vs non variable STR loci.
}
\description{
A function that takes a dataframe containing labeled STR loci and measure
variables, a number N and plots out the top N most common motifs and
the proportion of variable vs non variable STR loci.
}
\examples{
# Plot out the top 5 most common motifs
plotTopNCommonMotifs(df, 5)

}
\references{
R Core Team (2020). R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna,
Austria. URL https://www.R-project.org/.

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag
New York, 2016.
}
