% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTopNLeastVariableMotifs}
\alias{plotTopNLeastVariableMotifs}
\title{Plots the top N least variable or stable motifs from the dataframe.}
\usage{
plotTopNLeastVariableMotifs(df, num)
}
\arguments{
\item{df}{The dataframe created using labelSTR.R contaning labeled STR loci
and measurement variables.}

\item{num}{Specifies top how many least variable motifs to graph.}
}
\value{
Plots out a bar plot showing the top N least variable motifs.
}
\description{
A function that takes a dataframe containing labeled STR loci and measure
variables, a number N and plots out the top N least variable or stable motifs
}
\examples{
# Plot out the top 5 least variable motifs
data("ShortTandemRepeatsLoci")
head("ShortTandemRepeatsLoci")
mydf <- createVariables(df = ShortTandemRepeatsLoci, 0.1)
plotTopNMostVariableMotifs(mydf, 5)

}
\references{
R Core Team (2020). R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna,
Austria. URL https://www.R-project.org/.

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag
New York, 2016.
}
