% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelSTR.R
\name{readIntoDF}
\alias{readIntoDF}
\title{Load the text data with short tandem repeat (STR) loci into a dataframe.}
\usage{
readIntoDF(filepath)
}
\arguments{
\item{filepath}{The string filepath containing the text file.}
}
\value{
Returns a dataframe with columns \code{chr}, \code{start}, \code{end} and \code{motif}
}
\description{
A function that loads the text file with STR loci into a data frame. Each STR
locus has the following format:
\verb{chr<1-23/X/Y>:<START_POSITION>-<ENDPOSITION>:<REPEAT_UNIT/MOTIF>}.
For example, chr16:10000-11000:AT or chrX:900000-900100:TCTC.
These loci are separated by a tab space or new line.
}
\examples{
# Load STR loci data from str_data.txt located in the data folder of root dir
myDF <- readIntoDF("/data/str_data.txt")

}
\references{
R Core Team (2020). R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna,
Austria. URL https://www.R-project.org/.

Hadley Wickham and Jim Hester (2020). readr: Read Rectangular Text Data.
R package version 1.4.0.
https://CRAN.R-project.org/package=readr
}
