% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSTRExpansionAnalyzer.R
\name{runSTRExpansionAnalyzer}
\alias{runSTRExpansionAnalyzer}
\title{Launch Shiny App for STRExpansionAnalyzer}
\usage{
runSTRExpansionAnalyzer()
}
\value{
No return value.
}
\description{
Launch Shiny App for STRExpansionAnalyzer
}
\examples{
\dontrun{
runSTRExpansionAnalyzer()
}

}
\references{
Winston Chang, Joe Cheng, JJ Allaire, Yihui Xie and Jonathan McPherson
(2020). shiny: Web Application Framework for R. R package version 1.5.0.
https://CRAN.R-project.org/package=shiny
}
\author{
Michael Wong, \email{michaelz.wong@mail.utoronto.ca}
}
