% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSTRExpansionAnalyzer.R
\name{runSTRExpansionAnalyzer}
\alias{runSTRExpansionAnalyzer}
\title{Launch Shiny App for STRExpansionAnalyzer}
\usage{
runSTRExpansionAnalyzer()
}
\value{
No return value.
}
\description{
Launch Shiny App for STRExpansionAnalyzer
}
\examples{
\dontrun{
runSTRExpansionAnalyzer()
}

}
\author{
Michael Wong, \email{michaelz.wong@mail.utoronto.ca}
}
